Make issue card focusable for a11y services
The text on a warning card already has a custom content description set
for Talkback. Hence we need to mark the screenReaderFocusable attribute as "false" to avoid having a duplicate description.

But the importantForAccessibility attribute should still be set to true
so that it can be focused by other a11y services like Select to Speak,
hence removing the "no" values.

Bug: 266663122
Bug: 262871987
Test: Tested with "Select to speak"
Change-Id: I4871318a017d6f5e8f4d14763dac2fa0508734e2
Relnote: Fix a11y issue with warning cards