Open activities from notif actions properly
Previously, because the action's PendingIntent was not directly
starting the Activity, the notification shade would remain open. The
Activity was being opened, but it was hidden and this was poor UX.

Non-resolving actions don't need to be routed via the
SafetyCenterNotificationReceiver because they do not have the same
requirements as resolving actions (timeouts, setting "in-flight",
etc.)

The down-side of this approach is that we will lose the ability to log
an event when this kind of non-resolving action is clicked.

Bug: 266561097
Change-Id: I1bee46dd71f6d59033b611f4f83788499d61de52