Add codename check for VanillaIceCream
Since the new VIC codename has been added in ag/21274507, we need to have a codename check for V to ensure that T tests are not run on higher versions. Once the U version is no longer in pre-release, both of these checks can be removed. Tracking it in b/258228790.

Bug: 269792738
Test: atest CtsSafetyCenterTestCases SafetyCenterFunctionalTestCases
Change-Id: I728e8dd66804f43d2b5fefe4fd6cd00ed9d943f4