Pre-render current Safety Center Data.
This stops a lot of page jank, since PreferenceFragmentCompat expects
all preferences to be added to the PreferenceScreen by the end of
onCreatePreferences. Any subsequent changes are animated by the
recyclerview as additions, and any "static" content preconfigured in the
preference XML files is pushed around.

I initially thought we'd have to inflate or construct the screen
manaually (since PreferencFragmentCompat starts some "binding" process
with the recycler view when calling setPrefrencesFromResources(), but
it turns out its not necessary, the preferences just need to be added
before onCreatePreferences finishes.

Bug: 270333128
Test: Manual

Change-Id: Ie1c0830acb1eeebb7b82483244ef1d64261066e1