No need to kill app if app op was changed for modern apps.
This is a follow up to ag/13156539 in 2020 due to b/263324262#comment47.

We are already not killing upon granting app op for modern apps in
ag/13156539, and we don't need to kill for revoking app op for modern
apps either because the permission should have been revoked as well.

This also makes persistChanges() behave consistently with
revokeRuntimePermissions() when delayChanges = false.

Bug: 263324262
Test: presubmit
Change-Id: I7ab31bb25a05f5976aa5358d47ea684463e2fdad