Add and use Security & privacy icons
Includes one of the required new drawable icon assets, a helper
function to retrieve `Icon` instances from the SC resources and some
new logic to switch between the general/critical icons and the
associated colours (one for red warnings, another for all others).

There is an outstanding bug to add the general icon when it has been
signed off and currently the critical icon is used for all
notifications, but the logic and code is all present.

Also did a drive-by clean up of the resources tests to standardize
their names etc.

Test: Manually
Bug: 266052702
Change-Id: I491ec234a137970ae467d32d27125ce745299484