[Safety Labels] Separate hygiene and notification jobs
Currently both hygiene and notification logic run inside the
"runMainJob" function.

But, we realized we actually want to run two jobs, on different
schedules:

1. The first job is hygiene, which is a one-time job that runs after
   each boot (as soon as the device becomes idle).

2. The second job is for notifications. This should run periodically
   every 30 days, and the job schedule should persist across reboots.
   This job should also only run when the device is idle.

The main motivation for this is that the "App Data Sharing" screen also
depends on this data, and we need the hygiene logic to run after each
boot.

The notification job also runs hygiene, which may seem odd, but this is
necessary to avoid the edge case where the notification job runs
*before* the hygiene job. (This edge case is highly probable on first
boot, but extremely unlikely thereafter.)

Bug: 268685095
Test: atest PermissionControllerMockingTests:SafetyLabelChangesJobServiceTest
Test: atest CtsPermission3TestCases:SafetyLabelChangesJobServiceTest
Change-Id: I2f25d73fe13fb69a18d6bafd33cdce6118feaa53