Fix illustration width for device rotation
The IllustrationPreference from SettingsLib doesn't allow setting the image width greater than the image height. But to align with the other warning cards in Safety Center, we need the width to be greater than the height incase of device rotations / large-screen devices. Hence, changing to a custom implementation for the illustrations.

Note: It currently uses the placeholder images which might seem a bit expanded. These will be replaced with final assets from UX that are smaller in height.

Bug: 253171482
Test: atest SafetyCenterSubpagesTest
Change-Id: I1eef98708df96ef6c920cddf0d673bc537a2f02e