Log NOTIFICATION_POSTED atoms
Logs when a notification is posted or updated. Does not log when the
SC data changes and the notification is unchanged (but still present).

I have started writing automated tests for this in the child CL but
I'm running up against a problem there - will speak to @deweytyl to
get help with this later. In the meantime I was able to test manually
with statsd_testdrive.

Test: statsd_testdrive 473
Bug: 268309616
Change-Id: I0e73aa887337f9beb8899187412023be295f7c71