Make the StatsdLogger a static helper.
And remove the dependency on the config reader, which makes it thread
safe. Some logging code was calling it without holding the lock
so that should fix the potential data race.

This will also make it easier to test the logging code by allowing
logging in tests, and will allow disabling the code path entirely.

Bug: 265917109
Test: atest SafetyCenterTestCases
Change-Id: I8d591f3b6dd2fa3db7a03f8116656c1875c9ee3d