Implement "button groups" on issues in U.
This leaves the non-group behavior in place in T, since the button list
margin might be configured by an OEM (it's overlayable) and we wouldn't
want to use the button group style if they've stuck a thick margin in
there.

By only applying the new behavior in U, OEMs will be able to update
those values in their UDC release instead of the behavior automatically
triggering via mainline update.

Bug: 269725572
Test: atest CtsSafetyCenterTestCases SafetyCenterFunctionalTestCases
Change-Id: Iae86bf759989fe4a89e0aa07c50f6f89bfea4fa7