Continue passing package checks when overriding a real source.
If a test was overriding a "real" source then when the real package name
was trying to interact with our APIs it would fail to pass the package
checks. The reason for this is that the override took precedence
regardless of the calling package name, which isn't ideal as real
sources may continue to interact with our APIs during test.

This CL will allow real sources to continue interacting with our APIs in
tests, even if they're being overriden; but will no-op the calls. The
prior behavior only allowed this for real sources that weren't being
overridden in tests; so if a test config had a "real" source ID it would
create this issue.

Bug: 272333629
Test: atest CtsSafetyCenterTestCases
Change-Id: I2dd34360860150803033677fa92026e3573f29aa