Don't add mapping if no duplicates.
We're currently adding mapping to only those issue that have dedupKey,
so in other cases we still need to directly rely on the default group.
So no need to set mapping when there are no duplicates.

This should also help with not having to deal with useless extras or
empty extras in various tests.

Test: atest CtsSafetyCenterTestCases SafetyCenterFunctionalTestCases
Bug: 272456390
Change-Id: Iff1aff3e4a0021e1c715ab815761c5890a42ba00