[Safety Labels] Replace right/left with start/end
In this view, we have a few references to "right" and "left", which
resulted in an undesirable layout for RTL languages. Use "start" and
"end", instead, so that everything mirrors, including which side of our
elements we place padding.

Also, add textDirection="locale" and layoutDirection="locale" to the
root element. A vendor found that the text direction was incorrect for
an RTL language. This was unreproducable by us, so admittedly this
solution is somewhat of a guess (though it is recommmended as i18n best
practice, so shouldn't cause any harm).

Bug: 271422001
Test: atest CtsPermission3TestCases:android.permission3.cts.SafetyLabelChangesJobServiceTest
Change-Id: I30947281002545d8b9b209914e92a60f84661c96