Update issue styles to match new specs.
These changes apply to both T and U. Android TextViews don't render to
the exact size of the lineHeight (didn't realize this in T) so this CL
reduces some of the spacing between views to compensate for the extra
padding witin the TextViews.

This also fixes a bug with our pre-existing line heights, which weren't
being applied since they're not actually part of the TextAppearance
stylealbe and must be defined directly on views. Since these line
heights were never functional, I've decided to not make them
overlayable.

Fixes: 257973475
Fixes: 269731835
Test: atest CtsSafetyCenterTestCases SafetyCenterFunctionalTestCases
Change-Id: Ic121dc5010a50fef3084dca99bd022060fdfa2ac