Implement SourceState logging
Sets proper values for the new `SourceState` enum field in all
`SafetySourceStateCollected` atoms we log. To achieve I've added a new
source-to-state map to the `SafetySourceDataRepository` because it
didn't seem the pre-existing data was sufficient to fully disambiguate
all the new `SourceState` values, but I think that's fine and kind of
expected because the whole idea of this new enum is to unpack the
different underlying causes for "UNKNOWN" severities.

Fix: 26830917
Test: statsd_testdrive 471
Change-Id: Icb806c4603f9ce87a030672575894ba2ecd1524f