Define various writeSafetySourceStateCollected...
For the two linked bugs there will end up being two basic variations
of the existing `writeSafetySourceStateCollected` called in different
circumstances and with different sets of required params.

This CL is just to define these variations and the next one will make
use of the two new one (with the `...Automatic` variant being the same
as the single `writeSafetySourceStateCollected` that already exists
and thus already used here.

Test: N/A
Bug: 268309211
Bug: 268309213
Change-Id: I1896ec80875f49fba9e27ec321b295237e38eb46