Show the data of quiet mode entries.
But keep them disabled & their severity level as unspecified so that they
don't contribute to the overall status.

This is helpful to show the user some data if there is some, while
keeping the entry greyed out and the summary to "Work profile off" to
make it clear the provided data is stale and doesn't contribute to the
overall status.

This is particularly helpful for "hidden by default" entries which won't
show up unless some data is provided. In the case of quiet mode, it
would be good to continue showing these entries if we have some data for
them.

Bug: 267487752
Test: atest SafetyCenterFunctionalMultiUsersTestCases
Change-Id: I67f107cb361296e230bc72e6a1a8c6b84a236922