Use the explicit intent to open the quiet mode dialog.
My understanding is that the quiet mode dialog will redirect to the
original intent once quiet mode is turned off. Considering the implicit
intent restrictions on U+, it seems better to use the explicit intent
for this dialog as it has a higher chance of resolving once the user
turns the quiet mode off.

Bug: 266538628
Test: b/272493869
Change-Id: I8f36e28c554785809683e7dadc31d1b026fea6b6