Only override content description for first view in issue cards.
We'd like the screen reader to read "Alert." before reading out the rest
of the content of the issue card. Previously, we did this by using one
of two content description strings (one for issues with subtitle and one
for issues without subtitle).

Instead of creating additional strings for when attribution title is
present/absent, this CL just puts the special content description on the
first view in the card (either attribution title or issue title) and
modifies the issue card's a11y attributes to read out all the child
views.

Bug: 257972736
Test: atest SafetyCenterFunctionalTestCases
Change-Id: I6ef6f0026ff1d5576be167fa80485cf8be481fa5