Always use default pending intents for quiet mode.
Using the cached PendingIntents in this case ends up disabling the
paused work profile dialog. My assumption is that these cached
PendingIntents end up being cancelled as a side effect of pausing the
work profile, and as such triggering them causes an error rather than
deferring to the dialog.

Fix: 277737232
Test: manual, adding tests in b/272493869
Change-Id: I0b33630c5d60bb6a8e4f4e3b848c3db29de47095