Quiet mode dialog for hidden-by-default sources.
In quiet mode, if the source provides an intent use that if the default
pending intent cannot be generated.

This is to support some hidden-by-default Settings entries that handle
the work profile dialog on their own. For other sources, this may lead
to the entry being clickable but the dialog not showing up (as per
b/277737232). I'm not sure there is a straightforward way to check if
the PendingIntent is cancelled without triggering it?

Bug: 277873790
Test: manual, adding tests in b/272493869 (working on it, this really needs tests)
Change-Id: I97908f530769b6006dbb80a7d6f6e9e9ef630253