Ignore atoms from bg location source in tests
I haven't ascertained why this source is sending data so frequently
(on the ATH) that this is a problem, it's not behaving the same way
when running these tests locally.

But this should get the tests back into a stable state regardless and
we can investigate that as a secondary issue if necessary.

If I find that we need to make a third special intervention (to filter
out more atoms from prod sources) then at that point I will do
something more clever like filtering everything from the config or
whatever..

Fix: 278782808
Change-Id: Ie574ada3dc18a0347c7bd15a0618b445db42f62c