Fix activity controls for work showing in search without work profile.
"Activity controls for work" is a static entry within a non-static
group, so it's never surfaced as a SafetyCenterStaticEntry. As such, we
can actually remove it from search since it's technically surfaced as a
"dynamic" entry.

In general, the code was making the assumption that if a source was
static then it would be surfaced as a SafetyCenterStaticEntry. This is
not exactly right, as:

1. A dynamic entry can be surfaced a a SafetyCenterStaticEntry if it's
   within a static group (e.g. work policy info)
2. A static entry can be surfaced as a SafetyCenterEntry if it's within
   a dynamic group (e.g. activity controls)

Bug: 278860488
Test: manual
Change-Id: I8f2138a37c20cc54517d9c95ba04a978d027384f