Only use EN-US strings to lint config
When we delete strings, we tend to delete only the EN-US version and we
then rely on the translation pipeline to delete all translations. The
linter would have reported a false positive if a string is deleted
without deleting the reference in the config, because it would fallback
on a translation. (Note that the actual code on the device would not
fallback if the device is set to EN-US).

Also, improve the string checks by taking into consideration which SDK a
string is available from.

Bug: 257974388
Test: m out/soong/.intermediates/packages/modules/Permission/SafetyCenter/Resources/SafetyCenterResources/android_common/lint/lint-report.html
Test: atest ConfigLintCheckerTest
Change-Id: I7b8bfa432c77f06fd24b16d271d7c8a4a008d021