Do not allow source provided intents in quiet mode.
This intent is cached from when the profile wasn't in quiet mode, which
isn't ideal because the source won't be "refreshed" in the quiet mode.

We're seeing an issue with the Settings app where it provides the
"right" intent in the quiet mode (to open the dialog) — but because
we're not sending the broadcast we still have the old cached intent
which opens the page without the dialog (and probably it wouldn't
update even if we were sending the refresh broadcast due to the
profile being in the quiet mode).

This isn't a problem in most cases when the `PendingIntent` is
created for the profile, as the system adds that dialog when
attempting to send it. However, the Settings app sends that
`PendingIntent` in the profile parent as these screens do not support
being launched in the work profile (see ag/22714062).

This CL will make the Settings entry appear in the work profile, however
it will be greyed out (already the case) AND unclickable. Unlike most
other entries, it won't surface the quiet mode dialog, which is a
restriction that is preferable to allowing launching this intent in
quiet mode and making changes to a work profile that shouldn't be
meddled with.

Bug: 279573264
Test: manual
Change-Id: I4b4a1264d9ee5df17adef897e2c8e03e032d87eb