Remove unnecessary showNextRequest
showNextRequest() is already called by onRequestInfoLoad, so it's not
necessary to call here too. When it's called in onPermissionGrantResult
the same permission group is updated before any state has changed. On
Auto, this leads to the dialog being dismissed, showed again, and then
removed when the activity exits, causing a flickering effect.

Bug: 278507748
Test: atest PermissionTest23#testRequestPermissionFromTwoGroups
Change-Id: I1d9064c933c1555d6861504104737554faa783bd