DO NOT MERGE Correct sensor privacy state if sensor privacy not supported
The Safety Center QS page was incorrectly showing the camera and mic
toggles on devices that might not support them. This was fixed, but
anyone who had used those toggles before, now has no UI to re-enable mic
or camera, and is stuck with them disabled. This CL causes the
PermissionController to schedule a job, on boot, which will try to
correct the error

Bug: 280005257
Relnote: Fixes issue where users may have erroneously seen a sensor
privacy toggle, and corrects the state of the toggle
Test: manual

Change-Id: I5c53eea0bd581743d58bba7b8b552aa468b7a641