Fix Grant Permission Dialog numbering
The grant permission dialog shows a current and total page number.
It could happen that the current page number gets incremented by more
than 1 after proceeding from a page, resulting in an incorrect page
number that may greater than the total page number.

The number (derived from mCurrentRequestIdx) could be incremented
multiple times if showNextRequest() is called multiple times. That
method is called directly in the buttons' onClick handler, as well as
when the requestInfoLiveData updates as a result of the button press.

The showingNewGroup variable was meant to prevent incrementing the page
number if showNextRequest() is called for the same request. But the
equality check between the request message and current activity title
failed. The result of CharSequence#equals is undefined, and may fail due
to the objects being from different subclasses.

Fixing showingNextGroup by eg. using CharSequence#compare still doesn't
solve all the issues. The first time showNextRequest() is called, it
updates the UI with the mCurrentRequestIdx as page index, before
incrementing mCurrentRequestIdx. When showNextRequest() is called again,
the page is rendered again with the same requestInfo but a higher page
index. mCurrentRequestIdx doesn't get incremented again, but the
rendered page number ends up being off by one.

This change removes the mCurrentRequestIdx counter, and instead derives
the page number from the total page count and the remaining requestInfo
pages.

Relnote: Fixed page numbering in permission grant dialog.
Bug: 280755710
Test: manual
Change-Id: I4d44705cab18d36ed40c6be5f20744680dd13ea0