Fix some nullable warnings.
The `@NonNullByDefault` annotation wasn't working as expected, using a
fork from another team instead. Also added it to the notifications
package where it was missing.

Running null analysis on the code spawned a few warnings, I tried to fix
most / all of them here. Some are benign (the checker has to understand
that a certain variable cannot be null), but some were genuine issues.

Bug: 267476663
Test: atest CtsSafetyCenterTestCases
Change-Id: I992ed765d02c1a7b0a39997ea3718e1de09d4986