Use AndroidX nullable annotations.
Instead of the Android ones. It seems that the latter doesn't get along
well with the `@NonNullByDefault` annotation (it doesn't appear to be
taking precedence over it and causes a bunch of incorrect warnings in
IntelliJ).

The former, however, appears to be working as expected.
The Android annotation is typically used over the AndroidX one in system
server code because to avoid AndroidX dependencies in non-APK code.
But in our case, our code already depends on the AndroidX annotations so
there shouldn't be any harm in using them instead.

Bug: 259969737
Test: look at IntelliJ warnings
Change-Id: I058907ccf2cf069791e7c1d5b6a47efe67113b4e
Relnote: N/A