Use a theme attribute for the SC button theme.
We have been using a fixed material theme for Safety Center action
buttons, but this causes some problems when using a daynight theme to
make things work on the homepage, while Quick Settings has only a dark
theme (regardless of whether or not the device is in dark or light
mode).

Specifically, we had to set the secondary button text color to
textColorPrimaryInverse in light mode and textColorPrimary in dark mode.
Now we just use the appropriate dark-only MaterialComponents theme for
Quick Settings, and the DayNight theme for the full settings page.

Relnote: n/a
Test: manually verified
Change-Id: I5a92e12df1e4b887b25613e8711acb203a520cbf