Don't set errors on background refreshes timeout.
Background refreshes are more prone to timing out, clearing the
associated data and setting an error on each entry could end up leaving
the user in an unknown state.

This isn't much of an issue on T as the background refresh only happens
once, but on UDC the background refreshes are periodic so there is a
good chance of us clearing the state of the user when doing that.

AFAIK, this won't affect the logging and we will continue reporting the
sources as timed out in the pull atom callback, i.e:

Source data: <something, which is the uncleared cached data>
Source state: REFRESH_TIMEOUT
Timestamp: <last time the data was set, not when it timed out>

Bug: 282133501
Test: atest SafetyCenterFunctionalTestCases
Change-Id: I63f69f0b9e88fcc2526ad650a70a9244f4ed4896