Apply AndroidLockScreenFix earlier
For consistency with the SafetyEventFix and to ensure that we are
nowhere accidentally using the non-fixed incoming data.

This is a possibly-functional change if, accidentally, we are
somewhere (between where the fix was applied before and where it is
applied now) relying on the non-fixed data, but I think this is
unlikely!

Bug: 283780918
Relnote: N/A
Test: atest SafetyCenterManagerTest
Change-Id: Ia686eebf15eea3af61b84aadba6239a47a4f9314