Update description of v34 mic_toggle_description.
There should not be a full stop at the end of this sentence, which is
why it was updated on UDC (see ag/22354512).

However, some translations currently have a full stop at the end of
the sentence, causing the resources to get merged with the non-UDC
string.

Given this only happens in some locales, this causes resources
resolution issues. See b/287453782 for more details.

Bug: 285108009
Test: manual
Relnote: Updating description for better translations
Change-Id: Ic0a5157b7950581c8b3932f163f18b796bcad77e