Use a lock for SafetyCenterEnabledListener#mSafetyCenterEnabled.
Supposedly it was thread-safe already due to being called on
the same thread, but this setup is rather error-prone and relied
on the initial state being setup on the same thread as the one
updating the listener.

This thread is instantiated rather shortly after boot, so I wonder
if that's why the initial state wasn't the expected one in b/287898326.
It appears `#setInitialState` wasn't being called (or was called with
the wrong value).

Bug: 287898326
Test: atest CtsSafetyCenterTestCases
Relnote: N/A
Change-Id: Iee1d056a7d0d12dceab28ce4e45342e9eaee63ed