Add another SDK check.
I believe it's possible for the flag to be deleted.
In this case, #onPropertiesChanged documentation states
that the callback will still be called with the key in
`Properties`.

What I assume happens is that the value for the proprety
in this case happens to be `null` so the default value is
still applicable even if we checked that the value is in the
map prior to that.

Probably the flakyness we were seeing on TH was due to another
test clearing flags, causing this value to be toggled to `false`
when it should be `true`.

Bug: 287898326
Test: atest CtsSafetyCenterTestCases
Relnote: N/A
Change-Id: I432809d1823a558cc16598a7c52d05405161f4d1