Allow ROLE_SMS in managed profile conditionally
ROLE_SMS is only relevant in the managed profile if work profile telephony is enabled. If we enable ROLE_SMS in work profile with work profile telephony disabled, Work messaging app could access personal messages(since there is a single messaging database) and this gives user the perception that personal message data is leaking in to work profile.
Hence with this change ROLE_SMS would only be available in the work profile if work profile telephony is enabled.

Bug: 289197367
Test: atest android.devicepolicy.cts.telephony.WorkProfileTelephonyTest
atest android.devicepolicy.cts.ManagedSubscriptionsPolicyTest
Manual test(SMS role is not granted to OEM default messaging app when work profile telephony is not enabled)

Change-Id: Ia97fcb68140f537a65cfd465e1cb42799ea0f581