Include stopped packages for Safety Center user initiated broadcasts.
By default, stopped packages are excluded when sending a broadcast.
This is to avoid abuse for packages that wish to use system broadcasts
to get woken up.

This may cause issues in the case of Safety Center broadcasts,
as the user may open the UI while one of the target packages
is stopped.

In our case, all the packages we're targeting are known beforehand
(defined in a config file), so including stopped packages seems
adequate. We're also only doing that on user-initiated broadcasts
(page open or button click).

Bug: 283707833
Test: manual
Relnote: Refresh Safety Center for stopped packages too
Change-Id: Icb482fb4b6cdacd111f7b03df8911e1851126f15