Fix flakey interaction logging tests and remove @Ignore annotations.
openSubpageFromSettingsSearch_recordsEventWithSettingsNavigationSource
did not flake after removing the @Ignore annotation.

openSubpageFromHomepage_recordsEventWithSafetyCenterNavigationSource
always had one FULL view event and one SUBPAGE view event and variably a
second FULL view event recorded when going back from the subpage. I
added a sleep to see if the missing event was being recorded but not
propagated to the report but the tests still failed. Seems like
sometimes in the test navigation, the full page is destoryed/exited
before onStart is called and the event is logged.

Accordingly, I decided to narrow the assertion instead of messing with
the client-side code to ensure a second FULL event is always (or never)
recorded.

Relnote: n/a
Fixes: 278202773
Test: atest SafetyCenterHostSideTestCases
Change-Id: I30bd05d8372e0a68a2950dcb36c2439a5851fbc1