Replace UI clicks on notifications by assertions.
We can check if the notification has auto-cancel set on it,
without us having to rely on opening the tray and actually clicking
the notification.

The latter can be quite flaky because if notifications are collpased,
additional care needs to be taken to uncollapse them prior to clicking
them. UI testing in general is also less reliable.

Also added the UI specific rules, since the notifications are opening
the Safety Center UI at times, as a side effect of clicking these
notifications.

Bug: 292089550
Test: atest SafetyCenterNotificationTest
Relnote: N/A
Change-Id: I9d851974d5a1f7922d50985e45236db190f4438f