Relax/move some CTS tests on T.
Some of the requirements being tested in our CTS tests
were introduced in a fairly recent version of the mainline module.

As such, it may make more sense for these to be MTS requirements
instead of CTS; contsidering we allow running these tests on T.

Also fixed two other CTS to allow running on earlier versions
of mainline.

Bug: 292213555
Test: atest CtsSafetyCenterTestCases
Relnote: N/A
Change-Id: I686136f838971626cce5c7ec6e8dca4fa7d96ac7