Don’t create SC notification channels if notifications are disabled.
This doesn't take into account the flag being toggled, but we have
a separate bug to track that. This would probably require a DeviceConfig
listener to clear the notification state when flag goes from true
to false (and to create the channels when the opposite happens).

Bug: 287970155
Test: manual (not trivial to test because channels are also created when
posting notifications)
Relnote: N/A

Change-Id: I84d0681d40a6455136f488a5af94cb43d36056d1