Cleanup safety_center_show_error_entries_on_timeout flag.
This one is a bit trickier because tests rely on this in multiple ways:

- UI tests want to ensure that the data on the screen doesn't change,
however a refresh is triggered on page open and most tests set the data
manually rather than using the receiver to do so. The timeout could set
error on UI entries in the middle of the test, which isn't ideal. Worked
around this by setting the refresh timeout to the test timeout.

- Some existing tests were actively checking the behavior of that flag
and using the error listener as an easy way to check for timeouts.

I think it's a good thing to cleanup this flag regardless, as we ideally
want to have tests matching the production behavior, which wasn't the case
here (the flag was set to false in tests, but true in production code).

Bug: 289788052
Test: SafetyCenter tests
Relnote: N/A
Change-Id: I1f9dc8749d46f437c8dfbb98228ed1f042ee5821