Simplify waiting for UI to show/be gone
This keeps the current behavior of the tests, but (a) allows them to
complete much faster in success cases, and (b) adds TODOs for bugs that
need fixing.

This change makes the UI tests about 20% faster on average, although
fixes to the known bugs may remove some of this gain in future.

Bug: 292809502
Relnote: Speed improvement to CTS and MTS SafetyCenter tests
Test: atest SafetyCenterSubpagesTestCases SafetyCenterActivityTest
Change-Id: Ia891037f02e5019d0105002d4ce4342477bdca82