Update NoWifiStatePermissionTest to AndroidJUnit4
This test currently extends the deprecated AndroidTestCase, which has no
support for (1) Assume, or, (2) expected exceptions. Upgrade to
AndroidJUnit4, which gives us both features, and simplifies the tests.

Assume gives more informative test results when WiFi is not available
on a device, in which case we currently just show a passing test case.

Fix: 274613210
Test: atest NoWifiStatePermissionTest
Change-Id: Ie4f9216cc35c43723536144f92d1faee2343f2c3