Clean up temporary files in post-test runs
PermissionController has several test modules that run, and each pushes
temporary APK files to a corresponding directory. Two of them are:

- CtsPermissionMultiUserTestCases:
    - /data/local/tmp/cts/permissionmultiuser/...
- CtsPermissionUiTestCases
    - /data/local/tmp/cts/permissionui/...

Each module will create this corresponding directory with `mkdir -p
...`. But, these do not clean up their files when finished.

The result is that `/data/local/tmp/cts` may be created, and when it is
created, it will be created with whatever credentials are currently used
at the time.

If, later another test module tries to also push files into
`/data/local/tmp/cts/...`, it may run under different credentials, and
so may fail.

The solution used here is to clean up after use. It should be noted
though that, if this cleanup step were to somehow be prematurely
interrupted, then the cleanup would not happen, and the same issue would
occur again. Thus, a future solution might be to use paths like
`/data/local/tmp/cts-permission/`, `/data/local/tmp/cts-permissionui/`,
etc.

Bug: 291233073
Test: On go/abtd: run the :alltests test group
Change-Id: I99b0ae83cea4434ee76742179959b01ff7709bc8