Replace waitForIdle in CtsPermissionUiTestCases
We call uiAutomation.waitForIdle in many places in our tests, including
after every click. In most cases, the motivation for doing this was to
wait for an activity to start or finish, or to finish navigating to
another screen in the same activity.

But, waitForIdle has been called "a source of problems", because of its
unreliability for actually doing what we're trying to use it for, and
often leads to unpredictable timeouts, especially when the device is
under high load.

The suggested way to address this is to use UiAutomator's "wait for
transitions" methods: UiDevice.performActionAndWait(...),
UiObject2.clickAndWait(...), etc. These methods wait for specific
accessibility events.

The general approach in this change was to:

1. Remove any redundant waitForIdle calls (i.e., multiple waitForIdle
   calls in a row).
2. When opening a new window (i.e., bring a new activity to the
   foreground or invoke navigation), wrap the action with
   UiDevice.performActionAndWait(...). Then, any waitForIdle calls which
   follow can safely be removed.
3. Each time we click anything, if that click is expected to open a new
   window, use clickAndWait(Until.newWindow(), ...). Otherwise, simply
   use click() and don't bother waiting for anything. Then, any
   waitForIdle calls which follow can safely be removed.
4. In some cases, method arguments needed to be added (such as
   "waitForWindowTransition") which tell the method whether it should
   expect a new window event to occur or not after it occurs.

(After this, only a few waitForIdle calls remain. Either I haven't
been able to test those cases yet, or it's currently unclear how to
address them.)

Bug: 294875530
Test: atest CtsPermissionUiTestCases
Change-Id: I5b2a79dfa595a0291ccebd928873e93d1f12dec0