Fix ShellCommandPerimssionTest
The `ShellCommandPermission` tests previously checked `stderr` for a
line containing the string `Error:` to determine if the command failed
to run or not. Due to recent changes no output is generated when
checking whether the command can run, instead the process returns an
exit code of -1. This updates the permissions tests to check that the
process exits with an error code instead of relying on parsing the
output of stderr.

Bug: 288217526
Fixes: 293898385
Fixes: 289277264
Fixes: 293799054
Test: atest CtsPermissionTestCases:android.permission.cts.ShellCommandPermissionTest
Change-Id: I2b4fdc35582e5499859a8e8e189dca60ba84e3ce