Extract Bluetooth and Location test rules
Both NearbyDevicesPermissionTest and NearbyDevicesRenouncePermissionTest
ensure that Bluetooth and Location are enabled before running. (And, for
Bluetooth, they ensure that the adapter is running test mode.)

Since this behavior is common between both tests, and presumably future
tests might also want to use this behavior, it makes sense to extract
this into a Rule.

Test: atest NearbyDevicesPermissionTest NearbyDevicesRenouncePermissionTest
Bug: 296270764
Change-Id: Iace4c1e6840a2bc2983f1db640806b9beb458c21