Throw exception if wait-for-new-window times out
Recently, in CtsPermissionUiTestCases, in most places where we start an
activity, we surrounded the code with

        uiDevice.performActionAndWait({
            block()
        }, Until.newWindow(), NEW_WINDOW_TIMEOUT_MILLIS)

This will wait 20 seconds for an activity to start, before
continuing on.

One undesirable feature of performActionAndWait is that, if it times out
waiting, it will silently continue rather than throwing an exception.

It turns out, we can check its return value to determine whether the
timeout occurred or not. So, do so.

Bug: 298728114
Test: atest CtsPermissionUiTestCases
Change-Id: I5d4bd030b6fb56a8a0d5886ee8d0d60f1fb05972