[BackupHelper] Break the loop when END_DOCUMENT is reached
`skipToEndOfTag` does not consider the case END_DOCUMENT when calling
parser.next(); this causes an infinite loop. `skipToEndOfTag` is reached
because we catch an exception due to a corrupted file causing an
unexpected EOF and just continue to try to parse the file.

Relnote: Break out of loop in during restore at END_DOCUMENT
Bug: 299676312
Test: Push corrupted file force delayed restore by installing
Change-Id: I3eed0447cee8205c1db0d6b3f103938a4a4bdc07